Install Python dependencies early to benefit for Docker build cache
In order to benefit from Docker build cache, we install Python dependencies
separately from the scraper itself.

We do it early in the build pipe so that this layer is modified only if dependencies
are updated. Otherwise, only layers linked to scraper source code or zimui are updated,
leading to a much faster build + using less resources since we do not update a big layer
again and again for only a small scraper source code modification